+++ src/components/user-button.tsx
@@ -1,9 +1,6 @@
 "use client";
-import { roleCheck } from "@/lib/utils/role-check";
-import { Role } from "@prisma/client";
 import { User2 } from "lucide-react";
-import { signIn, signOut, useSession } from "next-auth/react";
-import Link from "next/link";
+import { signIn, signOut } from "next-auth/react";
 import { Avatar, AvatarFallback, AvatarImage } from "./ui/avatar";
 import { Button } from "./ui/button";
 import {
@@ -14,41 +11,33 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "./ui/dropdown-menu";
+import { Session } from "next-auth/types";
 
-export function UserButton({ role }: { role: Role }) {
-  const { data: session } = useSession();
-  const user = session?.user;
-
+export function UserButton({ session }: { session: Session | null }) {
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
         <Avatar className="cursor-pointer">
-          <AvatarImage src={user?.image ?? ""} />
+          <AvatarImage src={session?.user?.image ?? ""} />
           <AvatarFallback>
             <User2 />
           </AvatarFallback>
         </Avatar>
       </DropdownMenuTrigger>
       <DropdownMenuContent className="mt-3 w-60">
-        {user && (
+        {session?.user && (
           <>
             <DropdownMenuLabel className="py-4">
               <div className="flex flex-col space-y-1 pb-2.5">
-                <p className="text-sm font-medium leading-none">{user.name}</p>
+                <p className="text-sm font-medium leading-none">
+                  {session?.user.name}
+                </p>
                 <p className="text-xs leading-none text-muted-foreground">
-                  {user.email}
+                  {session?.user.email}
                 </p>
               </div>
               <DropdownMenuSeparator />
             </DropdownMenuLabel>
-
-            {roleCheck(role, [Role.STUDENT, Role.SUPERVISOR]) && (
-              <DropdownMenuItem>
-                <Link href="/account" className="w-full">
-                  <Button variant="link">My account</Button>
-                </Link>
-              </DropdownMenuItem>
-            )}
           </>
         )}
         <DropdownMenuItem>
